requirements.txt
azure-cli
numpy
azureml-sdk